requirements.txt
asgiref==3.8.1
Django==5.1.2
django-jazzmin==3.0.1
djangorestframework==3.15.2
python-decouple==3.8
PyYAML==6.0.2
sqlparse==0.5.1
ua-parser==0.18.0
user-agents==2.2.0
